function added for PUT and DELETE routes

+++ backend-nest-pradeo-test/src/apps/android-apps.service.ts
@@ -1,7 +1,9 @@
 
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Repository } from 'typeorm';
+import { NotFoundError } from 'rxjs';
+import { deleteAppFromHash } from 'src/utils/file.utils';
+import { DeleteResult, Repository } from 'typeorm';
 import { AndroidApp } from './android-app.entity';
 
 @Injectable()
@@ -9,18 +11,18 @@ export class AndroidAppsService {
   constructor(
     @InjectRepository(AndroidApp)
     private androidAppRepository: Repository<AndroidApp>,
-  ) {}
+  ) { }
 
   create(androidApp: AndroidApp): Promise<AndroidApp> {
     return this.androidAppRepository.save(androidApp);
   }
 
-  updateStatus(androidApp: AndroidApp): Promise<AndroidApp> {
+  update(androidApp: AndroidApp): Promise<AndroidApp> {
     return this.androidAppRepository.save(androidApp);
   }
 
   findAllByUser(userId: string): Promise<AndroidApp[]> {
-    return this.androidAppRepository.find({where: {user: userId}});
+    return this.androidAppRepository.find({ where: { user: userId } });
   }
 
   findAll(): Promise<AndroidApp[]> {
@@ -31,7 +33,18 @@ export class AndroidAppsService {
     return this.androidAppRepository.findOne(id);
   }
 
-  async remove(id: string): Promise<void> {
-    await this.androidAppRepository.delete(id);
+  async remove(id: string): Promise<DeleteResult> {
+
+    const appToDelete = await this.findOne(id);
+    if (appToDelete !== undefined && appToDelete !== null) {
+      const isFileDeleted = await deleteAppFromHash(appToDelete.hash);
+      if (isFileDeleted) {
+        console.log("File successfully deleted");
+      } else {
+        console.error("File not NotFoundError, deleting entity only");
+      }
+      return await this.androidAppRepository.delete(id);
+    }
   }
+
 }
